[resolves #1097] Collocate classloading integration tests under one module and package
