+++ itests/standalone/basic/src/main/java/org/wildfly/camel/test/classloading/CustomFailoverExceptionTest.java
