+++ itests/standalone/basic/src/main/java/org/wildfly/camel/test/classloading/ExplicitCamelDependencyTest.java
