+++ itests/standalone/basic/src/main/java/org/wildfly/camel/test/classloading/subB/FakeActivationSpec.java
@@ -17,16 +17,24 @@
  * limitations under the License.
  * #L%
  */
+package org.wildfly.camel.test.classloading.subB;
 
-package org.wildfly.camel.test.classloading.subA;
+import javax.resource.ResourceException;
+import javax.resource.spi.ActivationSpec;
+import javax.resource.spi.InvalidPropertyException;
+import javax.resource.spi.ResourceAdapter;
 
-import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.cdi.ContextName;
-
-@ContextName("cdi-context")
-public class CdiRouteBuilder extends RouteBuilder {
+public class FakeActivationSpec implements ActivationSpec {
     @Override
-    public void configure() throws Exception {
-        from("direct:start").log("Hello world");
+    public void validate() throws InvalidPropertyException {
+    }
+
+    @Override
+    public ResourceAdapter getResourceAdapter() {
+        return new FakeResourceAdapter();
+    }
+
+    @Override
+    public void setResourceAdapter(ResourceAdapter resourceAdapter) throws ResourceException {
     }
 }
