+++ itests/standalone/basic/src/main/resources/classloading/classpath-properties-context.xml
