+++ itests/standalone/basic/src/main/resources/classloading/custom-failover-context.xml
