+++ itests/standalone/basic/src/main/resources/classloading/test.properties
